requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
redis==4.5.5
pydantic==1.10.8
pynsq ==0.9.1